Fixed redirect bug

+++ frontend/app.js
@@ -15,12 +15,9 @@
             $('.button-collapse').sideNav('hide');
         }
 
+        // Event listener every route change
         $rootScope.$on('$routeChangeStart', (event, next, current) => {
-            let user = check_session($location.url());
-
-            if (!no_need_auth($location.url()) && typeof user === 'undefined') {
-                $location.path('/');
-            }
+            check_session($location.url());
         });
 
         // Routes that doesn't need authentication
@@ -33,34 +30,39 @@
             return no_auth_routes.indexOf(route) === -1 ? false : true;
         }
 
+        // check if the current route needs admin authentication (check if in the array)
         function need_admin_auth(route) {
             return routes_for_admin.indexOf(route) === -1 ? false : true;
         }
 
+        // check if the current route needs faculty user authentication (check if in the array)
         function need_faculty_user_auth(route) {
             return routes_for_faculty_user.indexOf(route) === -1 ? false : true;
-        }       
+        }
 
         // Check the backend for existing session
-        function check_session(next_route) {
+        function check_session(url) {
             let deferred = $q.defer();
 
             AuthenticationService
                 .GetUser()
                 .then((data) => {
-                    if (data.data === false) {
-                        localStorage.clear();
-                        $location.path('/');                      
+                    if (data.data === false && no_need_auth(url)) {
+                        $location.path(url);
                     }
-                    else {
+                    else if (data.data === false && !no_need_auth(url)) {
+                        localStorage.clear();
+                        $location.path('/');
+                    }
+                    else if (data.data !== false) {
                         localStorage.user = data.data;
                         if (data.data.role === 'Faculty User') {
-                            if (need_admin_auth(next_route) || next_route === '/') {
+                            if (need_admin_auth(url) || url === '/') {
                                 $location.path('/home');
                             }
                         }
                         else if (data.data.role === 'Administrator') {
-                            if (need_faculty_user_auth(next_route) || next_route === '/') {
+                            if (need_faculty_user_auth(url) || url === '/') {
                                 $location.path('/admin');
                             }
                         }
@@ -83,7 +85,7 @@
             })
             .when('/home', {
                 'controller'    :   'CourseCtrl',
-                'templateUrl'   :   'views/home.view.html' 
+                'templateUrl'   :   'views/home.view.html'
             })
             .when('/class', {
                 'controller'    :   'SectionCtrl',
